+++ buildconfig/Setup.SDL2.in
@@ -72,5 +72,5 @@ math src_c/math.c $(SDL) $(DEBUG)
 pixelcopy src_c/pixelcopy.c $(SDL) $(DEBUG)
 newbuffer src_c/newbuffer.c $(SDL) $(DEBUG)
 system src_c/system.c $(SDL) $(DEBUG)
-geometry src_c/geometry.c $(SDL) $(DEBUG)
+geometry src_c/geometry.c >$(SDL) $(DEBUG)
 window src_c/window.c $(SDL) $(DEBUG)
