+++ src_c/geometry_common.h
